athletes hooks. move useAuthAthlete hook to respective dir
