+++ src/hooks/services/athletes/useAuthenticatedAthlete.js
@@ -1,7 +1,7 @@
 import {useMemo} from 'react';
 
-import useBaseReadHook from './useBaseReadHook';
-import athletesApi from '../services/api/athletes';
+import useBaseReadHook from '../../useBaseReadHook';
+import athletesApi from '../../../services/api/athletes';
 
 function useAuthenticatedAthlete() {
